@@ -50,77 +50,80 @@
 		"animejs": "^3.0.0",
 		"antd": "3.15.0",
 		"classnames": "^2.2.6",
-		"echarts": "^4.3.0",
+		"color": "^3.1.2",
+		"echarts": "^4.7.0",
 		"fabric": "^3.5.1",
 		"gifler": "^0.1.0",
 		"i18next": "^19.0.0",
-		"i18next-browser-languagedetector": "^2.2.3",
+		"i18next-browser-languagedetector": "^2.2.4",
 		"lodash": "^4.17.10",
 		"mediaelement": "^4.2.9",
 		"prop-types": "^15.6.2",
-		"react": "^16.4.1",
-		"react-ace": "^6.1.4",
+		"react": "^16.13.1",
+		"react-ace": "^8.1.0",
 		"react-color": "^2.14.1",
 		"react-custom-scrollbars": "^4.2.1",
-		"react-dom": "^16.4.1",
+		"react-dom": "^16.13.1",
 		"react-dom-factories": "^1.0.2",
 		"react-helmet": "^5.2.0",
-		"react-hot-loader": "^4.12.12",
+		"react-hot-loader": "^4.12.20",
 		"react-json-view": "^1.19.1",
 		"resize-observer-polyfill": "^1.5.1",
 		"store": "^2.0.12",
-		"typescript": "^3.5.3",
-		"uuid": "^3.3.3",
+		"typescript": "^3.8.3",
+		"uuid": "^7.0.3",
 		"warning": "^4.0.3"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.2.2",
-		"@babel/plugin-proposal-class-properties": "^7.2.3",
-		"@babel/plugin-proposal-decorators": "^7.2.3",
-		"@babel/plugin-proposal-object-rest-spread": "^7.2.0",
-		"@babel/plugin-syntax-dynamic-import": "^7.2.0",
-		"@babel/plugin-transform-runtime": "^7.2.0",
-		"@babel/polyfill": "^7.2.5",
-		"@babel/preset-env": "^7.2.3",
-		"@babel/preset-react": "^7.0.0",
-		"@babel/preset-typescript": "^7.1.0",
+		"@babel/core": "^7.9.0",
+		"@babel/plugin-proposal-class-properties": "^7.8.3",
+		"@babel/plugin-proposal-decorators": "^7.8.3",
+		"@babel/plugin-proposal-object-rest-spread": "^7.9.5",
+		"@babel/plugin-syntax-dynamic-import": "^7.8.3",
+		"@babel/plugin-transform-runtime": "^7.9.0",
+		"@babel/polyfill": "^7.8.7",
+		"@babel/preset-env": "^7.9.5",
+		"@babel/preset-react": "^7.9.4",
+		"@babel/preset-typescript": "^7.9.0",
 		"@types/animejs": "^3.1.0",
-		"@types/echarts": "^4.1.15",
-		"@types/fabric": "^3.5.0",
-		"@types/lodash": "^4.14.144",
-		"@types/react": "^16.7.18",
-		"@types/react-dom": "^16.0.11",
-		"@types/uuid": "^3.4.6",
+		"@types/color": "^3.0.1",
+		"@types/echarts": "^4.4.6",
+		"@types/fabric": "^3.6.2",
+		"@types/lodash": "^4.14.150",
+		"@types/react": "^16.9.34",
+		"@types/react-dom": "^16.9.6",
+		"@types/react-helmet": "^5.0.15",
+		"@types/uuid": "^7.0.2",
 		"@types/warning": "^3.0.0",
-		"babel-eslint": "^10.0.1",
-		"babel-loader": "^8.0.5",
+		"babel-eslint": "^10.1.0",
+		"babel-loader": "^8.1.0",
 		"babel-plugin-dynamic-import-webpack": "^1.1.0",
-		"babel-plugin-import": "^1.11.0",
-		"css-loader": "^2.1.0",
+		"babel-plugin-import": "^1.13.0",
+		"css-loader": "^3.5.2",
 		"del": "^5.1.0",
-		"eslint": "^5.12.0",
+		"eslint": "^6.8.0",
 		"eslint-config-airbnb": "^17.1.0",
-		"eslint-loader": "^2.1.1",
-		"eslint-plugin-import": "^2.14.0",
-		"eslint-plugin-jsx-a11y": "^6.1.2",
-		"eslint-plugin-react": "^7.12.3",
-		"file-loader": "^3.0.1",
-		"fs-extra": "^8.1.0",
+		"eslint-loader": "^4.0.0",
+		"eslint-plugin-import": "^2.20.2",
+		"eslint-plugin-jsx-a11y": "^6.2.3",
+		"eslint-plugin-react": "^7.19.0",
+		"file-loader": "^6.0.0",
+		"fs-extra": "^9.0.0",
 		"html-webpack-plugin": "^3.2.0",
 		"less": "^2.7.3",
-		"less-loader": "^4.1.0",
+		"less-loader": "^5.0.0",
 		"path": "^0.12.7",
-		"serve": "^11.2.0",
-		"style-loader": "^0.23.1",
-		"terser-webpack-plugin": "^1.2.3",
-		"tslint": "^5.12.1",
-		"tslint-react": "^3.6.0",
-		"typedoc": "^0.15.4",
-		"url-loader": "^1.1.2",
-		"webpack": "^4.28.4",
-		"webpack-cli": "^3.2.1",
-		"webpack-dev-server": "^3.1.14",
-		"webpack-merge": "^4.2.1",
-		"workbox-webpack-plugin": "^3.6.3"
+		"serve": "^11.3.0",
+		"style-loader": "^1.1.4",
+		"terser-webpack-plugin": "^2.3.5",
+		"tslint": "^6.1.1",
+		"tslint-react": "^5.0.0",
+		"typedoc": "^0.17.4",
+		"url-loader": "^4.1.0",
+		"webpack": "^4.43.0",
+		"webpack-cli": "^3.3.11",
+		"webpack-dev-server": "^3.10.3",
+		"webpack-merge": "^4.2.2",
+		"workbox-webpack-plugin": "^5.1.2"
 	}
 }
