@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Collapse, notification, Input, message } from 'antd';
-import uuid from 'uuid/v4';
+import { v4 } from 'uuid';
 import classnames from 'classnames';
 import i18n from 'i18next';
 
@@ -110,7 +110,7 @@ class ImageMapItems extends Component {
                 message.info('Already drawing');
                 return;
             }
-            const id = uuid();
+            const id = v4();
             const option = Object.assign({}, item.option, { id });
             if (item.option.type === 'svg' && item.type === 'default') {
                 this.handlers.onSVGModalVisible(item.option);
@@ -120,7 +120,7 @@ class ImageMapItems extends Component {
         },
         onAddSVG: (option, centered) => {
             const { canvasRef } = this.props;
-            canvasRef.handler.add({ ...option, type: 'svg', id: uuid(), name: 'New SVG' }, centered);
+            canvasRef.handler.add({ ...option, type: 'svg', id: v4(), name: 'New SVG' }, centered);
             this.handlers.onSVGModalVisible();
         },
         onDrawingItem: (item) => {
@@ -208,7 +208,7 @@ class ImageMapItems extends Component {
             if (dt.types.length && dt.types[0] === 'Files') {
                 const { files } = dt;
                 Array.from(files).forEach((file) => {
-                    file.uid = uuid();
+                    file.uid = v4();
                     const { type } = file;
                     if (type === 'image/png' || type === 'image/jpeg' || type === 'image/jpg') {
                         const item = {
